Moves to tdyutils.h to private directory
