+++ include/private/tdyutils.h
