+++ src/tdympfao.c
@@ -1,6 +1,6 @@
 #include <private/tdycoreimpl.h>
 #include <private/tdymeshimpl.h>
-#include "tdyutils.h"
+#include <private/tdyutils.h>
 #include <private/tdymemoryimpl.h>
 #include <petscblaslapack.h>
